remove quant install in docker building

+++ script/install.sh
@@ -29,27 +29,31 @@ bash install_mpi4py.sh
 bash install_pytorch.sh pypi
 popd
 
+build_quant=0
+if [[ $build_quant -eq 1 ]]; then
+   # Obtain the python version from the system.
+   python_version=$(python3 --version 2>&1 | awk '{print $2}' | awk -F. '{print $1$2}')
+   # Download and install the AMMO package from the DevZone.
+   wget https://developer.nvidia.com/downloads/assets/cuda/files/nvidia-ammo/nvidia_ammo-0.5.0.tar.gz
+   tar -xzf nvidia_ammo-0.5.0.tar.gz
+   pip install nvidia_ammo-0.5.0/nvidia_ammo-0.5.0-cp$python_version-cp$python_version-linux_x86_64.whl
+   rm -rf $llmtk/nvidia_ammo*
 
-
-# Obtain the python version from the system.
-python_version=$(python3 --version 2>&1 | awk '{print $2}' | awk -F. '{print $1$2}')
-# Download and install the AMMO package from the DevZone.
-wget https://developer.nvidia.com/downloads/assets/cuda/files/nvidia-ammo/nvidia_ammo-0.5.0.tar.gz
-tar -xzf nvidia_ammo-0.5.0.tar.gz
-pip install nvidia_ammo-0.5.0/nvidia_ammo-0.5.0-cp$python_version-cp$python_version-linux_x86_64.whl
-rm -rf $llmtk/nvidia_ammo*
-
-# Install the additional requirements
-pushd $root/quantization
-pip install cython # extra dependancy
-pip install -r requirements.txt
-popd
+   # Install the additional requirements
+   pushd $root/quantization
+   pip install cython # extra dependancy
+   pip install -r requirements.txt
+   popd
+fi
 
 # for mistral model running
 pip install --upgrade flash-attn
 # for mistral model building
 pip install pynvml
 
+# for dgtrt
+pip install pybind11
+pip install pybind11-stubgen
 rm -f ~/.wgetrc
 
 popd
